Replace deprecated GetVersionEx
MSDN: GetVersionEx may be altered or unavailable for releases after Windows 8.1. With the release of Windows 8.1, the behavior of the GetVersionEx API has changed in the value it will return for the operating system version. The value returned by the GetVersionEx function now depends on how the application is manifested.